checking and adding buildings to state

+++ src/Building.cpp
@@ -12,10 +12,10 @@ int buildingsDist(const Building & b1, const Building & b2) {
     const vector<vector<char>> & b2Plan = b2.getProject()->getPlan();
 
     int dist = INT_MAX;
-    for (int row1 = b1Y; row1 < b1Y + b1Plan.size(); row1++)
-        for (int col1 = b1X; col1 < b1X + b1Plan[0].size(); col1++)
-            for (int row2 = b2Y; row2 < b2Y + b2Plan.size(); row2++)
-                for (int col2 = b2X; col2 < b2X + b2Plan[0].size(); col2++)
+    for (size_t row1 = b1Y; row1 < b1Y + b1Plan.size(); row1++)
+        for (size_t col1 = b1X; col1 < b1X + b1Plan[0].size(); col1++)
+            for (size_t row2 = b2Y; row2 < b2Y + b2Plan.size(); row2++)
+                for (size_t col2 = b2X; col2 < b2X + b2Plan[0].size(); col2++)
                     dist = min(dist, manhattanDistance(row1,col1,row2,col2));
     return dist;
 }
